Update phoenix file structure
