+++ config/dev.exs
@@ -6,6 +6,14 @@ use Mix.Config
 # The watchers configuration can be used to run external
 # watchers to your application. For example, we use it
 # with brunch.io to recompile .js and .css sources.
+config :berlim, BerlimWeb.Endpoint,
+  http: [port: 4000],
+  debug_errors: true,
+  code_reloader: true,
+  check_origin: false,
+  watchers: []
+
+# ## SSL Support
 #
 # In order to use HTTPS in development, a self-signed
 # certificate can be generated by running the following
@@ -13,28 +21,22 @@ use Mix.Config
 #
 #     openssl req -new -newkey rsa:4096 -days 365 -nodes -x509 -subj "/C=US/ST=Denial/L=Springfield/O=Dis/CN=www.example.com" -keyout priv/server.key -out priv/server.pem
 #
-# The `http:` config below can be replaced with:
-# https: [port: 4000, keyfile: "priv/server.key", certfile: "priv/server.pem"],
+# The `http:` config above can be replaced with:
+#
+#     https: [port: 4000, keyfile: "priv/server.key", certfile: "priv/server.pem"],
 #
 # If desired, both `http:` and `https:` keys can be
 # configured to run both http and https servers on
 # different ports.
-config :berlim, Berlim.Endpoint,
-  http: [port: 4000],
-  debug_errors: true,
-  code_reloader: true,
-  check_origin: false,
-  watchers: []
-
 
 # Watch static and templates for browser reloading.
-config :berlim, Berlim.Endpoint,
+config :berlim, BerlimWeb.Endpoint,
   live_reload: [
     patterns: [
       ~r{priv/static/.*(js|css|png|jpeg|jpg|gif|svg)$},
       ~r{priv/gettext/.*(po)$},
-      ~r{web/views/.*(ex)$},
-      ~r{web/templates/.*(eex)$}
+      ~r{lib/berlim_web/views/.*(ex)$},
+      ~r{lib/berlim_web/templates/.*(eex)$}
     ]
   ]
