+++ config/prod.exs
@@ -1,18 +1,20 @@
 use Mix.Config
 
-# For production, we configure the host to read the PORT
-# from the system environment. Therefore, you will need
-# to set PORT=80 before running your server.
+# For production, we often load configuration from external
+# sources, such as your system environment. For this reason,
+# you won't find the :http configuration below, but set inside
+# BerlimWeb.Endpoint.init/2 when load_from_system_env is
+# true. Any dynamic configuration should be done there.
 #
-# You should also configure the url host to something
-# meaningful, we use this information when generating URLs.
+# Don't forget to configure the url host to something meaningful,
+# Phoenix uses this information when generating URLs.
 #
 # Finally, we also include the path to a cache manifest
 # containing the digested version of static files. This
-# manifest is generated by the mix phoenix.digest task
+# manifest is generated by the mix phx.digest task
 # which you typically run after static files are built.
-config :berlim, Berlim.Endpoint,
-  http: [:inet6, port: {:system, "PORT"}],
+config :berlim, BerlimWeb.Endpoint,
+  load_from_system_env: true,
   url: [host: "example.com", port: 80],
   cache_static_manifest: "priv/static/cache_manifest.json"
 
@@ -24,7 +26,7 @@ config :logger, level: :info
 # To get SSL working, you will need to add the `https` key
 # to the previous section and set your `:url` port to 443:
 #
-#     config :berlim, Berlim.Endpoint,
+#     config :berlim, BerlimWeb.Endpoint,
 #       ...
 #       url: [host: "example.com", port: 443],
 #       https: [:inet6,
@@ -39,7 +41,7 @@ config :logger, level: :info
 # We also recommend setting `force_ssl`, ensuring no data is
 # ever sent via http, always redirecting to https:
 #
-#     config :berlim, Berlim.Endpoint,
+#     config :berlim, BerlimWeb.Endpoint,
 #       force_ssl: [hsts: true]
 #
 # Check `Plug.SSL` for all available options in `force_ssl`.
@@ -54,7 +56,7 @@ config :logger, level: :info
 # Alternatively, you can configure exactly which server to
 # start per endpoint:
 #
-#     config :berlim, Berlim.Endpoint, server: true
+#     config :berlim, BerlimWeb.Endpoint, server: true
 #
 
 # Finally import the config/prod.secret.exs
