+++ lib/berlim/application.ex
@@ -1,0 +1,31 @@
+defmodule Berlim.Application do
+  use Application
+
+  # See https://hexdocs.pm/elixir/Application.html
+  # for more information on OTP Applications
+  def start(_type, _args) do
+    import Supervisor.Spec
+
+    # Define workers and child supervisors to be supervised
+    children = [
+      # Start the Ecto repository
+      supervisor(Berlim.Repo, []),
+      # Start the endpoint when the application starts
+      supervisor(BerlimWeb.Endpoint, []),
+      # Start your own worker by calling: Berlim.Worker.start_link(arg1, arg2, arg3)
+      # worker(Berlim.Worker, [arg1, arg2, arg3]),
+    ]
+
+    # See https://hexdocs.pm/elixir/Supervisor.html
+    # for other strategies and supported options
+    opts = [strategy: :one_for_one, name: Berlim.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  def config_change(changed, _new, removed) do
+    BerlimWeb.Endpoint.config_change(changed, removed)
+    :ok
+  end
+end
