+++ lib/berlim/repo.ex
@@ -1,3 +1,11 @@
 defmodule Berlim.Repo do
   use Ecto.Repo, otp_app: :berlim
+
+  @doc """
+  Dynamically loads the repository url from the
+  DATABASE_URL environment variable.
+  """
+  def init(_, opts) do
+    {:ok, Keyword.put(opts, :url, System.get_env("DATABASE_URL"))}
+  end
 end
