+++ lib/berlim_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule BerlimWeb.PageController do
+  use BerlimWeb, :controller
+
+  def index(conn, _params) do
+    render conn, "index.html"
+  end
+end
