+++ lib/berlim_web/router.ex
@@ -1,0 +1,26 @@
+defmodule BerlimWeb.Router do
+  use BerlimWeb, :router
+
+  pipeline :browser do
+    plug :accepts, ["html"]
+    plug :fetch_session
+    plug :fetch_flash
+    plug :protect_from_forgery
+    plug :put_secure_browser_headers
+  end
+
+  pipeline :api do
+    plug :accepts, ["json"]
+  end
+
+  scope "/", BerlimWeb do
+    pipe_through :browser # Use the default browser stack
+
+    get "/", PageController, :index
+  end
+
+  # Other scopes may use custom stacks.
+  # scope "/api", BerlimWeb do
+  #   pipe_through :api
+  # end
+end
