+++ lib/berlim_web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule BerlimWeb.LayoutView do
+  use BerlimWeb, :view
+end
