+++ lib/berlim_web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule BerlimWeb.PageView do
+  use BerlimWeb, :view
+end
