+++ test/berlim_web/controllers/page_controller_test.exs
@@ -1,0 +1,8 @@
+defmodule BerlimWeb.PageControllerTest do
+  use BerlimWeb.ConnCase
+
+  test "GET /", %{conn: conn} do
+    conn = get conn, "/"
+    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
+  end
+end
