+++ test/berlim_web/views/layout_view_test.exs
@@ -1,0 +1,3 @@
+defmodule BerlimWeb.LayoutViewTest do
+  use BerlimWeb.ConnCase, async: true
+end
