+++ test/berlim_web/views/page_view_test.exs
@@ -1,0 +1,3 @@
+defmodule BerlimWeb.PageViewTest do
+  use BerlimWeb.ConnCase, async: true
+end
