--- test/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule Berlim.PageControllerTest do
-  use Berlim.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get conn, "/"
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
-  end
-end
