+++ test/support/channel_case.ex
@@ -1,11 +1,11 @@
-defmodule Berlim.ChannelCase do
+defmodule BerlimWeb.ChannelCase do
   @moduledoc """
   This module defines the test case to be used by
   channel tests.
 
   Such tests rely on `Phoenix.ChannelTest` and also
   import other functionality to make it easier
-  to build and query models.
+  to build common datastructures and query the data layer.
 
   Finally, if the test case interacts with the database,
   it cannot be async. For this reason, every test runs
@@ -20,24 +20,18 @@ defmodule Berlim.ChannelCase do
       # Import conveniences for testing with channels
       use Phoenix.ChannelTest
 
-      alias Berlim.Repo
-      import Ecto
-      import Ecto.Changeset
-      import Ecto.Query
-
-
       # The default endpoint for testing
-      @endpoint Berlim.Endpoint
+      @endpoint BerlimWeb.Endpoint
     end
   end
+
 
   setup tags do
     :ok = Ecto.Adapters.SQL.Sandbox.checkout(Berlim.Repo)
-
     unless tags[:async] do
       Ecto.Adapters.SQL.Sandbox.mode(Berlim.Repo, {:shared, self()})
     end
-
     :ok
   end
+
 end
