+++ test/support/conn_case.ex
@@ -1,11 +1,11 @@
-defmodule Berlim.ConnCase do
+defmodule BerlimWeb.ConnCase do
   @moduledoc """
   This module defines the test case to be used by
   tests that require setting up a connection.
 
   Such tests rely on `Phoenix.ConnTest` and also
   import other functionality to make it easier
-  to build and query models.
+  to build common datastructures and query the data layer.
 
   Finally, if the test case interacts with the database,
   it cannot be async. For this reason, every test runs
@@ -19,26 +19,20 @@ defmodule Berlim.ConnCase do
     quote do
       # Import conveniences for testing with connections
       use Phoenix.ConnTest
-
-      alias Berlim.Repo
-      import Ecto
-      import Ecto.Changeset
-      import Ecto.Query
-
-      import Berlim.Router.Helpers
+      import BerlimWeb.Router.Helpers
 
       # The default endpoint for testing
-      @endpoint Berlim.Endpoint
+      @endpoint BerlimWeb.Endpoint
     end
   end
+
 
   setup tags do
     :ok = Ecto.Adapters.SQL.Sandbox.checkout(Berlim.Repo)
-
     unless tags[:async] do
       Ecto.Adapters.SQL.Sandbox.mode(Berlim.Repo, {:shared, self()})
     end
-
     {:ok, conn: Phoenix.ConnTest.build_conn()}
   end
+
 end
