+++ test/test_helper.exs
@@ -1,4 +1,4 @@
-ExUnit.start
+ExUnit.start()
 
 Ecto.Adapters.SQL.Sandbox.mode(Berlim.Repo, :manual)
