--- test/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule Berlim.LayoutViewTest do
-  use Berlim.ConnCase, async: true
-end
