--- test/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule Berlim.PageViewTest do
-  use Berlim.ConnCase, async: true
-end
