--- web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule Berlim.LayoutView do
-  use Berlim.Web, :view
-end
