--- web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule Berlim.PageView do
-  use Berlim.Web, :view
-end
